requirements.txt
attrdict==2.0.1
bokeh==3.0.3
#   contourpy==1.0.6
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
graphviz==0.20.1
Jinja2==3.1.2
kiwisolver==1.4.4
#   MarkupSafe==2.1.1
MarkupSafe==2.1.2
#   matplotlib==3.6.2
matplotlib==3.6.3
packaging==23.0
#   pandas==1.5.2
pandas==1.5.3
Pillow==9.4.0
pip==22.3.1
# pkg-resources==0.0.0
pydot==1.4.2
pyparsing==3.0.9
python-dateutil==2.8.2
#   pytz==2022.7
pytz==2022.7.1
PyYAML==6.0
scipy==1.10.0
#   setuptools==65.7.0
setuptools==66.1.1
six==1.16.0
tornado==6.2
wxPython==4.2.0
xyzservices==2022.9.0

#   protobuf==3.20.1
protobuf==4.21.12
grpcio~=1.51.1
grpcio-tools~=1.51.1
jsonpickle==3.0.1
numpy==1.24.1
#   networkx~=2.5
networkx~=3.0
#typing-extension
pylightning~=0.0.7.3
#   requests==2.28.1
requests==2.28.2
google~=3.0.0
gcloud
#   google-api-python-client==2.72.0
google-api-python-client==2.73.0

pyln-client
